Removed the `GTID` field from the `GaleraState` struct
As we're still not sure what to do on the operator side with the GTID we decided to not export it in the `GaleraState`. It would also force us to import the `GaleraState` struct to the operator repository so we chose to just not export it for the moment instead.

https://mariadb-operator.slack.com/archives/C056RAECH0W/p1699990598511999?thread_ts=1699350363.009529&cid=C056RAECH0W